backend/requirements.txt
# --- Core Web ---
fastapi==0.111.0
uvicorn[standard]==0.30.1

# --- Pydantic, Jinja (HTML/Schemas) ---
pydantic>=2.8.0
pydantic-core>=2.18.2
jinja2==3.1.4

# --- Databases ---
pymongo==4.7.2
redis[hiredis]==5.0.3                    # or redis[hiredis]==5.1.0b1 for async Python 3.12+

# --- Authentication & Security ---
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# --- ML & Text Processing ---
scikit-learn>=1.5
joblib>=1.4
nltk>=3.9
pandas

# --- Logging & Observability ---
structlog==24.2.0
loguru==0.7.2
python-json-logger>=2.0                  # optional, if using json logs

# --- API Rate Limiting ---
slowapi==0.1.9

# --- Python Version Constraint ---
# python_version >= "3.12" and < "3.13"

# --- Testing and QA Tools (optional, recommended for CI/CD) ---
# pytest==8.2.1
# pytest-cov==5.0.0
# httpx==0.27.0
# black==24.4.2
# ruff==0.4.6

# --- SSL/TLS (recommended for production) ---
# cryptography
# certifi
